Move mobile_authenticate tests to a separate file
